permissions and user create delete update done

+++ app/Http/Controllers/RoleController.php
@@ -3,21 +3,37 @@
 namespace App\Http\Controllers;
 
 use App\Helpers\ResponseHelper;
+use App\Services\RoleService;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
-use Spatie\Permission\Models\Permission;
+use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Validator;
 
 class RoleController extends Controller
 {
+    private $roleService;
+    function __construct(){
+
+        // Add middleware for route protection
+        $this->middleware('auth');
+        $this->middleware('can:view users')->only(['index', 'show']);
+        $this->middleware('can:create users')->only(['store']);
+        $this->middleware('can:edit users')->only(['update']);
+        $this->middleware('can:delete users')->only(['destroy']);
+        $this->roleService = new RoleService();
+    }
+
+
     /**
      * Display a listing of roles.
      */
     public function index(Request $request)
     {
+        // Fetch all roles with their associated permissions
         $roles = Role::with('permissions')->get();
 
-        return ResponseHelper::success('', $roles);
+        // Return the formatted roles in the response
+        return ResponseHelper::success('Roles are fetched.', $this->roleService->formattedData($roles));
     }
 
     /**
@@ -32,7 +48,7 @@ class RoleController extends Controller
         ]);
 
         if ($validator->fails()) {
-            return ResponseHelper::error('Validation faild!', $validator->errors());
+            return ResponseHelper::error('Validation failed!', $validator->errors());
         }
 
         $role = Role::create(['name' => $request->name]);
@@ -40,7 +56,7 @@ class RoleController extends Controller
         if ($request->has('permissions')) {
             $role->syncPermissions($request->permissions);
         }
-        return ResponseHelper::success('Role created successfully', $role);
+        return ResponseHelper::success('Role created successfully', $this->roleService->formattedSingleData($role));
     }
 
     /**
@@ -48,8 +64,9 @@ class RoleController extends Controller
      */
     public function show($id)
     {
-        $role = Role::with('permissions')->findOrFail($id);
-        return ResponseHelper::success('', $role);
+        $role = Role::with('permissions')->find($id);
+        if(!$role) return ResponseHelper::error('Role not found!', []);
+        return ResponseHelper::success('Role data', $this->roleService->formattedSingleData($role));
     }
 
     /**
@@ -57,7 +74,8 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $role = Role::findOrFail($id);
+        $role = Role::find($id);
+        if(!$role) return ResponseHelper::error('Role not found!', []);
 
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|unique:roles,name,' . $id,
@@ -74,7 +92,7 @@ class RoleController extends Controller
         if ($request->has('permissions')) {
             $role->syncPermissions($request->permissions);
         }
-        return ResponseHelper::success('Role updated successfully', $role);
+        return ResponseHelper::success('Role updated successfully', $this->roleService->formattedSingleData($role));
     }
 
     /**
@@ -82,9 +100,10 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
-        $role = Role::findOrFail($id);
+        $role = Role::find($id);
+        if(!$role) return ResponseHelper::error('Role not found!', []);
+        $roleCopyData=clone $role;
         $role->delete();
-
-        return response()->json(['message' => 'Role deleted successfully']);
+        return ResponseHelper::success('Role deleted successfully', $this->roleService->formattedSingleData($roleCopyData));
     }
 }
